.on('error') handler needs to be set right after net.createConnection is called.
Previouisly, if ImapConnection.connect was called with a non-existent hostname, an error would be raised before the error handler was set, and would percolate to the top level as an unhandled exception.

(Q: Is there ever any guarantee that the error handler will be set before an error is raised? It's not clear to me how node could guarantee this, but it seems like something one would like to guarantee.)