remove support for partial body fetching
We have no way to match up a partial fetch response if less than the number of
octets we requested were returned, due to the way IMAP responds to partial fetches.

Example: We request BODY[TEXT]<0.32>, but BODY[TEXT] is only 16 bytes,
         then we get back: BODY[TEXT]<0> {16}
         This leaves us with no way to find out what the original
         length request was.

Because of this and the fact that the server can return requested
values in any order, I am disabling partial fetches entirely.