Header parsing fails for headers with no value.
The following header breaks the header parsing:

    X-Face:
     H@&[wkk?l:Zx:8i_5b....

There is no whitespace after the colon, so it doesn't match the header
regex. This causes an error since there is no toLowerCase method on
m[1], which is undefined.

Added a second regex which only gets executed in the rare cases that
reHdr fails. Should not impact any mail parsing that was working
previously not should it impact the performance.