Parse box names with square brackets and no quotes correctly
`parseBoxList` did not correctly parse box names that have brackets and
do not have quotes. For example

`'(\\HasChildren \\Noselect) "/" [Airmail]/To Do'`, would return the
name `Airmail`, instead of `[Airmail]/To Do`. This is because `parseExpr` parses
anything within [] or () as a list, unless it's wrapped in quotes.

Given that `parseExpr` is used in many other cases, I did not want to
break this assumption and unknowingly break something else. Instead, I
updates `parseBoxList` to wrap the name in quotes before passing it to
`parseExpr` if it does not have quotes.

This commit also adds tests for several of these cases

(also add a .gitignore)