Potential fix for error with unhandled ECONNRESET
This error simply means that the other side closed the connection in a way that was probably not normal.

An example is: a socket connection can be closed by the other party abruptly due to several reasons and you will see this error / exception on your side.

In this case we need close our socket too. Before this fix I experienced the following behaviour:
- everything works great
- after long time this error occured, probably it can be connected with logout from gmail or network error, I do not know
- application could detect error, but does not reset connection
- only restart of all system allows to establish connection again

Now I want to listen in my application on connection close and end, then reset connection in my own code. I invite to discuss if it is good solution, because of I feel it is ok, but do not know how to test it.

https://github.com/mscdex/node-imap/issues/710